\name{BigBioSeqData-package}
\alias{BigBioSeqData-package}
\alias{BigBioSeqData}
\docType{package}
\title{
\packageTitle{BigBioSeqData}
}
\description{
\packageDescription{BigBioSeqData}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BigBioSeqData}
\packageIndices{BigBioSeqData}
Resources for BioC 2016 workshop, entitled, "Managing big biological sequence data with Biostrings and DECIPHER".}
\author{
\packageAuthor{BigBioSeqData}

Maintainer: \packageMaintainer{BigBioSeqData}
}
\keyword{ package }
\seealso{
}
\examples{
}
